Rewrite Spot marker to Spot icon
